Stop tracking .uvoptx file: (Empty commit):
I ran
  git update-index --assume-unchanged ./path/to/freertos1.uvoptx

This keeps the currently committed version of the .uvoptx file (which
is good because it has some important config information that should
stay with the project), but one won't have to constantly commit
changes when I change the window configuration or add a breakpoint,
etc.  These are meaningless changes that should NOT be tracked.